+++ openfl/federated/task/runner_fe.py
@@ -4,7 +4,7 @@
 
 import numpy as np
 
-from openfl.component.collaborator.collaborator import OptTreatment
+from openfl.utilities.enum_types import OptTreatment
 from openfl.utilities import split_tensor_dict_for_holdouts
 from openfl.utilities import TensorKey
 from .runner import TaskRunner
