+++ openfl/federated/task/runner_keras.py
@@ -11,8 +11,8 @@ from warnings import simplefilter
 
 import numpy as np
 
-from openfl.component.collaborator.collaborator import OptTreatment
 from openfl.utilities import Metric
+from openfl.utilities.enum_types import OptTreatment
 from openfl.utilities import split_tensor_dict_for_holdouts
 from openfl.utilities import TensorKey
 from .runner import TaskRunner
