+++ openfl/federated/task/runner_pt.py
@@ -12,8 +12,8 @@ import torch as pt
 import torch.nn as nn
 import tqdm
 
-from openfl.component.collaborator.collaborator import OptTreatment
 from openfl.utilities import Metric
+from openfl.utilities.enum_types import OptTreatment
 from openfl.utilities import split_tensor_dict_for_holdouts
 from openfl.utilities import TensorKey
 from .runner import TaskRunner
